requirements.txt
# Core data stack
pandas>=2.2
numpy>=1.26
polars>=1.0

# Visualization
matplotlib>=3.9
seaborn>=0.13
plotly>=5.22

# Statistical analysis
scipy>=1.13
statsmodels>=0.14

# Data loading (multi-format)
openpyxl>=3.1          # Excel
pyarrow>=16.0          # Parquet, Arrow
sqlalchemy>=2.0        # SQL databases
requests>=2.32         # API calls

# Jupyter
jupyter>=1.0
ipywidgets>=8.1

# Backend
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27

# AI / LLM
openai>=1.40

# Utilities
tqdm>=4.66
python-dotenv>=1.0
rich>=13.7             # Pretty terminal output
